% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{colMedians}
\alias{colMedians}
\title{calcualtes the median of each column}
\usage{
colMedians(x)
}
\arguments{
\item{infile}{matrix}
}
\value{
the median of each column in a vector
}
\description{
calcualtes the median of each column
}
