% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{get.close}
\alias{get.close}
\title{Gets the sample closest to a day}
\usage{
get.close(map, day)
}
\arguments{
\item{infile}{map and day as integer}
}
\description{
Gets the sample closest to a day
}
