% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{norm}
\alias{norm}
\title{Normalizes metabolite matrix with metaboanalyst}
\usage{
norm(samples.pick, metabolites)
}
\arguments{
\item{infile}{list of samples being used and matrix of raw metabolite concentrations}
}
\value{
glog normalized metabolite concentrations
}
\description{
Normalizes metabolite matrix with metaboanalyst
}
