% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{paired.map}
\alias{paired.map}
\title{Generate a map with only paired samples}
\usage{
paired.map(map, pats)
}
\arguments{
\item{infile}{original map and list of patients}
}
\value{
map with random pairs
}
\description{
Samples are paired randomly sith sample function.
}
