% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{paired.violin.beta}
\alias{paired.violin.beta}
\title{make violin plot and connects pairs}
\usage{
paired.violin.beta(map, var, plot.each = F)
}
\value{
the number of pairs
}
\description{
make violin plot and connects pairs
}
