% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{pairwiseRF}
\alias{pairwiseRF}
\title{Use Random forest analysis compare two groups}
\usage{
pairwiseRF(
  mat,
  map,
  dis_1 = "T 0",
  dis_2 = "B 7",
  num_of_col = 50,
  Accuray.cutoff = 0,
  Factor = 1,
  plot = T
)
}
\arguments{
\item{infile}{mat = matrix, map is a dataframe that must have a description column}
}
\value{
mean decrease in accuracy when each feature in the matrix is excluded from the RF
}
\description{
Use Random forest analysis compare two groups
}
