% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.22.06.22.R
\name{plot.meta}
\alias{plot.meta}
\title{Plots PCoA of metabolite  euclidean dissimilarity and colors points  by group}
\usage{
\method{plot}{meta}(meta.pick, map.pick, return.points = F)
}
\arguments{
\item{infile}{map and metabolite matrix}
}
\value{
PERMANOVA result
}
\description{
Plots PCoA of metabolite  euclidean dissimilarity and colors points  by group
}
