% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R
\name{remove_rare}
\alias{remove_rare}
\title{Remove rare features from matrix}
\usage{
remove_rare(table, cutoff_pro)
}
\value{
matrix with feature below the cutoff removed
}
\description{
Remove rare features from matrix
}
