% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.22.06.18.R,
%   R/mice_functions.22.06.22.R
\name{norm}
\alias{norm}
\title{Normalizes metabolite matrix with metaboanalyst}
\usage{
norm(samples.pick, metabolites)

norm(samples.pick, metabolites)
}
\arguments{
\item{infile}{samples included and metabolite matrix}
}
\value{
glog normalized metabolite concentrations

normalized matrix of samples in samples.pick vector
}
\description{
Normalizes metabolite matrix with metaboanalyst

Normalizes mouse metabolite matrix with Metaboanalyst
}
